+++ lib/onlyoffice_webdriver_wrapper/webdriver.rb
@@ -217,9 +217,9 @@ module OnlyofficeWebdriverWrapper
     def drag_and_drop(xpath, x1, y1, x2, y2, mouse_release: true)
       canvas = get_element(xpath)
       if mouse_release
-        @driver.action.move_to(canvas, x1, y1).click_and_hold.move_by(x2, y2).release.perform
+        @driver.action.move_to(canvas, x1.to_i, y1.to_i).click_and_hold.move_by(x2, y2).release.perform
       else
-        @driver.action.move_to(canvas, x1, y1).click_and_hold.move_by(x2, y2).perform
+        @driver.action.move_to(canvas, x1.to_i, y1.to_i).click_and_hold.move_by(x2, y2).perform
       end
     rescue ArgumentError
       raise "Replace 'click_and_hold(element)' to 'click_and_hold(element = nil)' in action_builder.rb"
@@ -282,7 +282,7 @@ module OnlyofficeWebdriverWrapper
     end
 
     def left_mouse_click(xpath, x_coord, y_coord)
-      @driver.action.move_to(get_element(xpath), x_coord, y_coord).click.perform
+      @driver.action.move_to(get_element(xpath), x_coord.to_i, y_coord.to_i).click.perform
     end
 
     # Context click on locator
@@ -303,7 +303,7 @@ module OnlyofficeWebdriverWrapper
       if browser == :firefox
         element.send_keys %i[shift f10]
       else
-        @driver.action.move_to(element, x_coord, y_coord).context_click.perform
+        @driver.action.move_to(element, x_coord.to_i, y_coord.to_i).context_click.perform
       end
     end
 
@@ -323,15 +323,15 @@ module OnlyofficeWebdriverWrapper
     def click_on_locator_coordinates(xpath_name, right_by, down_by)
       wait_until_element_visible(xpath_name)
       element = @driver.find_element(:xpath, xpath_name)
-      @driver.action.move_to(element, right_by, down_by).perform
-      @driver.action.move_to(element, right_by, down_by).click.perform
+      @driver.action.move_to(element, right_by.to_i, down_by.to_i).perform
+      @driver.action.move_to(element, right_by.to_i, down_by.to_i).click.perform
     end
 
     def right_click_on_locator_coordinates(xpath_name, right_by = nil, down_by = nil)
       wait_until_element_visible(xpath_name)
       element = @driver.find_element(:xpath, xpath_name)
-      @driver.action.move_to(element, right_by, down_by).perform
-      @driver.action.move_to(element, right_by, down_by).context_click.perform
+      @driver.action.move_to(element, right_by.to_i, down_by.to_i).perform
+      @driver.action.move_to(element, right_by.to_i, down_by.to_i).context_click.perform
     end
 
     def double_click(xpath_name)
@@ -341,13 +341,13 @@ module OnlyofficeWebdriverWrapper
 
     def double_click_on_locator_coordinates(xpath_name, right_by, down_by)
       wait_until_element_visible(xpath_name)
-      @driver.action.move_to(@driver.find_element(:xpath, xpath_name), right_by, down_by).double_click.perform
+      @driver.action.move_to(@driver.find_element(:xpath, xpath_name), right_by.to_i, down_by.to_i).double_click.perform
     end
 
     def action_on_locator_coordinates(xpath_name, right_by, down_by, action = :click, times = 1)
       wait_until_element_visible(xpath_name)
       element = @driver.find_element(:xpath, xpath_name)
-      (0...times).inject(@driver.action.move_to(element, right_by, down_by)) { |acc, _elem| acc.send(action) }.perform
+      (0...times).inject(@driver.action.move_to(element, right_by.to_i, down_by.to_i)) { |acc, _elem| acc.send(action) }.perform
     end
 
     def click_on_one_of_several_by_text(xpath_several_elements, text_to_click)
@@ -450,7 +450,7 @@ module OnlyofficeWebdriverWrapper
 
     def mouse_over(xpath_name, x_coordinate = 0, y_coordinate = 0)
       wait_until_element_present(xpath_name)
-      @driver.action.move_to(@driver.find_element(:xpath, xpath_name), x_coordinate, y_coordinate).perform
+      @driver.action.move_to(@driver.find_element(:xpath, xpath_name), x_coordinate.to_i, y_coordinate.to_i).perform
     end
 
     def element_present?(xpath_name)