+++ Changelog.md
@@ -12,3 +12,4 @@
 ### Fixes
 * Fix `Webdriver#alert_exists?` for `firefox`
 * Do not crash on `Webdrvier#browser_logs` in `firefox`
+* Fix crashes on `@driver.action.move_to` non-integer coordinates